refactor EnvironmentControl and buttons
